pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gemini-model-router"
version = "0.1.0"
description = "Local-first agentic terminal that routes prompts across local Gemma 4, Gemini CLI, and Claude Code."
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "pydantic>=2.7",
    "pyyaml>=6.0",
    "sentence-transformers>=3.0",
    "numpy>=1.26",
    "prompt_toolkit>=3.0",
    "halo>=0.0.31",
    "ulid-py>=1.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.6",
]
configsite = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
]
tuning = [
    "transformers>=4.46",
    "peft>=0.13",
    "trl>=0.12",
    "bitsandbytes>=0.44",
    "accelerate>=1.0",
    "datasets>=3.0",
    "torch>=2.4",
]

[project.scripts]
router = "router.cli:main"
router-repl = "router.repl:main"
router-eval = "router.eval.review:main"
router-config = "router.configsite.server:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "smoke: end-to-end CLI tests that hit live backends (opt-in via -m smoke)",
]
# Default run skips smoke; use `pytest -m smoke` to opt in.
addopts = "-m 'not smoke'"

[tool.ruff]
line-length = 100
target-version = "py311"
extend-exclude = [".venv", "paperbanana", "experiments/figures"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["B011"]
